fix some Category and Address API

+++ src/main/java/com/example/eCommerceUdemy/service/AddressServiceImpl.java
@@ -1,15 +1,19 @@
 package com.example.eCommerceUdemy.service;
 
+import com.example.eCommerceUdemy.exception.APIException;
 import com.example.eCommerceUdemy.exception.ResourceNotFoundException;
 import com.example.eCommerceUdemy.model.Address;
 import com.example.eCommerceUdemy.model.User;
 import com.example.eCommerceUdemy.payload.AddressDTO;
 import com.example.eCommerceUdemy.repository.AddressRepository;
+import com.example.eCommerceUdemy.repository.OrderRepository;
 import com.example.eCommerceUdemy.repository.UserRepository;
 import com.example.eCommerceUdemy.util.AuthUtil;
 import jakarta.transaction.Transactional;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -26,6 +30,8 @@ public class AddressServiceImpl implements AddressService {
     private AuthUtil authUtil;
     @Autowired
     private UserRepository userRepository;
+    @Autowired
+    private OrderRepository orderRepository;
 
     @Override
     public AddressDTO createAddress(AddressDTO addressDTO, User user) {
@@ -97,15 +103,24 @@ public class AddressServiceImpl implements AddressService {
     public String deletedAddress(Long addressId) {
         Address addressFromDB = addressRepository.findById(addressId)
                 .orElseThrow(() -> new ResourceNotFoundException("Address", "id", addressId));
+        // check Address link to Order
+        boolean isUsedInOrder = orderRepository.existsByShippingAddress_AddressId(addressId);
+        if(isUsedInOrder){
+            throw new APIException(HttpStatus.NOT_ACCEPTABLE, "Address is used in order");
+        }
 
         //update address for one user
-        User user = addressFromDB.getUser();
-        user.getAddresses().removeIf(address -> address.getAddressId().equals(addressId));
-        userRepository.save(user);
+        try {
+            User user = addressFromDB.getUser();
+            user.getAddresses().removeIf(address -> address.getAddressId().equals(addressId));
+            userRepository.save(user);
 
-        addressRepository.delete(addressFromDB);
+            addressRepository.delete(addressFromDB);
 
-        return "Address removed with id: " + addressId;
+            return "Address removed with id: " + addressId;
+        } catch (Exception e) {
+            throw new APIException(HttpStatus.BAD_REQUEST, "Something went wrong");
+        }
     }
